🚚 ユーザ登録POSTエンドポイントを移動
